requirements.txt
fastapi
uvicorn[standard]
deta
pymongo
pydantic-settings